fix(ui): DOMA-8365 moved progressIndicator steps to separate file
